Add scaling support for 8-bit types
- Also clean up some code
- Add new autotune config
